requirements.txt
Django==4.1.3
anthropic
requests
django-cors-headers
djangorestframework
python-dotenv
